requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.3.0
python-dotenv==1.0.1
typing_extensions==4.15.0
Werkzeug==3.1.5
